Add a script for testing upgrades.
This introduces a new script, ./contrib/internal/test-upgrade.py, which will
test upgrading Review Board installs. It can be given a list of branch or
tag names, and will install a brand new installation using the first version,
and then perform an upgrade on every subsequent version.

By default, this will use sqlite as the database, but it takes
--database-type, --database-name, --database-user, and --database-password
parameters in order to test with other databases.

This also updates prepare-dev to take the same parameters, plus --no-db
and --no-media.