Disallow duplicate draft comment replies and fix file attachment replies.
The API was allowing multiple draft comment replies by a user, and due to
a design flaw in the web UI, we were attempting to create a reply every time
the user re-edited a reply.

The API will look to see if a draft comment already exists. If it does, it
will update it and send a HTTP 303 See Other, just like we do with draft
reviews.

While the web UI is still technically doing the wrong thing, the result is that
it's fixed, and we can tackle rewriting that part of code later on.

Along with this, file attachment comment replies have been fixed. They were
completely non-functioning and untested before. They've now been updated to
work and unit tests were added to make sure it stays working.

Fixes bugs #2154 and #2173