Remove the evolution history for add/remove_diff_revision.
An evolution for the diff revisions were added in 1.6 RC2, and was then
removed due to some issues in the migration. It was important to have
the remove_diff_revision history for anyone working on the developer install,
but it should be safe to remove now.

My hope is that this will fix a bug some people are hitting where they can't
upgrade to RC2. I'm unable to figure out the exact cause of this, but this
change is suspect, as it's the only evolution-related change that took place
between RC1 and 2.