Add a cache backend setting to the admin UI.
This introduces a new setting in the admin UI for changing the configured
cache server. This is one of the major settings from settings_local.py that
people may still need to change.

Reviewed at http://reviews.reviewboard.org/r/2450
This introduces a new setting in the admin UI for changing the configured
cache server. This is one of the major settings from settings_local.py that
people may still need to change.

Reviewed at http://reviews.reviewboard.org/r/2450
This introduces a new setting in the admin UI for changing the configured
cache server. This is one of the major settings from settings_local.py that
people may still need to change.

Reviewed at http://reviews.reviewboard.org/r/2450
This introduces a new setting in the admin UI for changing the configured
cache server. This is one of the major settings from settings_local.py that
people may still need to change.

Reviewed at http://reviews.reviewboard.org/r/2450/