Keep e-mails from breaking when the sender has no e-mail address.
Our e-mail code made the assumption that the sender would have an e-mail
address, and would add it to the list of recipients. In some setups, this
user may be a bot with no actual address to list. Django would fail when it
saw an entry that didn't have a "@" in it. Now we're more careful about adding
this entry.