Edit the perforce+stunnel docs a bit to make the version requirements clear.
We require stunnel version 3 (which allows you to specify everything on the
command-line) instead of version 4 (which requires a config file). Make this
clear in the docs, as well as being a little bit more specific about hosts,
ports, and the order in which things need to be set up.