Add documentation on access control.
This covers the various settings for group and repository access control
on the Review Groups and Repository configuration pages, respectively. It also
provides a general Access Control page in the Configuration section, which
summarizes the options, points to the proper settings, and gives a summary of
what conditions are required for a review request to be accessible to a user.

Fixes bug #2203