Remove the reports feature.
The reports feature is one of the best kept secrets in Review Board. We
never exposed it, so you had to know the URL. It was pretty basic, too.

It has now been removed. The equivalent functionality now lives in the
RBReports extension, which will probably be improved in time.