Remove old Django/Django-Evolution cruft from manage.py.
In order to work around some old model changes around Django 1.0 time,
we monkey-patched Django Evolution a bit. That's unnecessary these days,
and it's been long enough, so remove that.

We also don't explicitly check for Django in manage.py anymore. We know
we're using Django at this point, and clearly the check for 1.0 wasn't
helping us anywhere. In practice, these checks aren't important anymore
for a production install, and developers should just run prepare-dev to
get the deps we need.