Fix captions for newly-added screenshots and files.
The way that screenshot captions (and file attachment captions) worked was still
a little bit broken. When creating a screenshot via the "Add Screenshot" form,
it would set caption = draft_caption = user entered text, and everything work
work fine when the draft was published. When using the DnD upload, however, it
just attaches the item and then expects the user to change the caption in the
thumbnail. Setting the caption there will set draft_caption, but because it was
a newly added screenshot, we'd skip over it when copying draft_caption ->
caption during the publish phase.