Fix comment turds.
If you click "Add comment", start writing something, and then change your mind,
it's only natural to cancel the comment addition. The comment will be cancelled,
but it leaves a turd on the review list with your name and an edit pencil.

This change hooks up the inline editor cancel event to check if the initial
value was empty, and if so, remove the comment item.