Add a way to "archive" repository names before deleting them.
When deleting a repository through the API, it doesn't actually delete it, it
just changes the "visible" flag. If someone wants to re-use a repository name
(such as "main"), they'll then get errors about it already being taken.

This change adds an "archive_name" field which can be sent to PUT
/api/repositories/<id>/. When this field is True, we'll automatically assign a
new name to the repository which shouldn't conflict with anything a user might
choose.