Fix tabs to spaces and a KeyError in the issue summary change.
The previous change inadvertantly introduced some tabs to the code. This
converts those back to spaces, cleans up a little bit of style, and also
fixes a KeyError when change descriptions were in the review.

While I was in there, I added a **kwargs to get_all_comments and made
the queries filter on that, allowing us to specifically get comments
with an issue open.