Order group names in the dashboard by name.
The order in which groups under the "My Groups" or "Watched Groups" section in
the dashboard sidebar would appear was reliant on iterating a python dict, which
meant essentially random. Since lists like this are important to be able to
scan, we should sort them.

Based on David Stanek's change at http://reviews.reviewboard.org/r/2571/