Fix up bits of style for the admin UI, and simplify some logic.
This reworks the new admin UI code to be compliant with our style
guidelines for Python, HTML, JavaScript, and some of the docs.

It also simplifies some of the code that was doing too much
unnecessarily.

The next step will be to create a more solid concept of Widgets, and
make expand/collapse settings store in cookies.