Remove the fade-in of the new admin UI and improve loading overall.
The new admin UI faded in on load. This was kind of neat but also a bit
overkill, especially after the first couple times. This seems to have
been partially there to work around an issue involving collapsed states
for widgets.

If a widget started collapsed, we were temporarily hiding it, giving
flot time to draw (it won't draw in a display: none, which the FAQ even
points out), then showing it. This was happening during the fade-in
time.

I've worked around this by triggering an event telling widgets when
they've been shown. We can use this to draw the first time we get the
signal when flot is involved. This allows us to just initially show the
widgets, default the collapsed ones to actually being collapsed.

Along with this, I've changed the collapse/expand animation to slide the
content, rather than fade in/out.