Move our stylesheets to lessCSS.
This updates all of our stylesheets to be .less files. In the process,
I've cleaned up most of the rules to add some structure and, in some
cases, added definitions for colors and other values.

There's still a long ways to go here, but it's a good start and
definitely makes things more manageable.

Where it made sense, I renamed some classes in the HTML to make things
more clear and organized.

This will work perfectly fine on production installs or developer installs.