Hard-code our jQuery and jQuery UI requirements for 1.6.
The versions of jQuery and jQuery UI will be updated in Djblets soon,
but Review Board 1.6 isn't designed to work with those changes yet.
Bundle custom jQuery templates to force the versions we need.

We also now bundle the versions of jQuery specifically. Djblets will no
longer be hosting this, and it'll be up to consumers of Djblets to
supply the jQuery if they don't wish to use the Google-hosted versions.