Switch to using django-pipeline and change our versioning.
This moves from django-compress, which is deprecated, to
django-pipeline. We only do pipeline work when we're on a production
system, or DEBUG is off. We fall back to using the less runtime on a
non-production DEBUG setup.

Rather than using MEDIA_SERIAL, we're now using Pipeline to version the
generated files. This will slightly reduce startup time, since we won't
have to stat the media files, and it also means we can version the files
individually instead of as a group. We'll have more fine-grained caching
as a result.

A future change will remove our use of MEDIA_SERIAL.