Fix the import path for the fallback PIL Image.
It seems easy for PIL to be installed wrong. When it is installed wrong,
you need to insert Image instead of PIL.Image. Our unit tests attempted
this, but failed, as we tried to import Image from Image. Clearly,
nobody has hit this before now, but I have on my new deployment. This is
now fixed.