Fix SPF compliance.
Currently Reviewboard sends emails with a Return-Path of the current
user. This is not SPF compliant, and as a result review emails will be
rejected if a user's domain implements SPF with a reject all failed
messages policy and a reciever checks SPF.

This corrects this by setting the From header directly instead, and
passing the site default from email address as the Return-Path address.
This is needed on the KDE Reviewboard instance to ensure email delivery
runs well.

Reviewed at http://http://reviews.reviewboard.org/r/2702/