Add support for per-resource mimetypes in the API.
Resources in the API are now returned with per-resource mimetypes. These
allow a client to identify the resource and make an intelligent decision
about what that payload represents without having to hard-code anything
about the location of the resource. An example would be client bindings
that could not only generically handle any resources but could also
provide utility functions for specific ones.

Unit tests have been updated to verify each and every resource we
provide.

This change will be made available for 1.5 and 1.6, so that we could
take advantage of this in RBTools.