Raise NotImplementedError, not NotImplemented.
We weren't being consistent with returning NotImplementedError vs.
NotImplemented. This is a bit confusing as Python has both, but the
latter is intended as a return value for operators rather than an
exception.

This never seemed to have bit us in practice, but in theory we'd fail to
catch exceptions in some cases, so fix it.