Fix checking for file existence when the file is empty on Git.
Git was using the file contents as an indicator as to whether a file
exists when using raw file URLs. However, this fails on empty files,
such as a standard __init__.py.

What we really care about is whether we can successfully fetch a file,
or whether we'll get an HTTP error. So, ignore the contents, and assume
the file exists if there's no error.