Fix potential KeyErrors when submitting the repo form without a path.
If the repository form in the admin is submitted with a custom hosting service
and an empty path, our validation would try to access the key directly without
checking if it exists. This is fine for most fields but not those which come
from ModelForm, so it fails. This change adds some simple validation to check
that the path is specified.