Fix DefaultReviewer.objects.for_repository for non-repo LocalSite reviewers.
DefaultReviewer objects worked properly with regards to LocalSite rooting when
they had a repository associated, but not when they were global. This change
adds an additional parameter to for_repository which will make sure that only
LocalSite-specific default reviewers will be matched. This also includes a unit
test for the same.

Fixes bug 2367.