Avoid crashing if an scmtool returns bogus text during validation.
When an scmtool returns text with an unexpected encoding as an exception (such
as if pysvn can't load the DLL because something is broken, and the system
locale is Big-5), we would spew a traceback. This catches text encoding errors
and forces it to "replace" instead, which means that users will see a bunch of
boxes with question marks in them, but at least it's better than a traceback,
and (in this particular case) includes the text "DLL load failed".

Tested by manually raising an exception with the bogus text and verifying the
displayed error in the UI.

Fixes bug 1921.