Add validation for the cache backend setting.
If users type the wrong thing into the cache backend setting in the admin site,
it can save to the database, which will then cause any request to the site to
generate a traceback. This is particularly annoying because settings are stored
as a big JSON blob, so it's not even easy to fix.

This change adds validation using django's parse_backend_uri--if the form can
parse it, we know that this won't cause problems later on.

Fixes bug 2347.