Convert our HTTP requests from urllib2 to requests.
In addition to being a downright awful API to use, urllib2 has a fundamental
flaw in that it will only send HTTP Basic Auth information if the first request
for the resource comes back with a www-authenticate header. This doesn't work
for many HTTPS mercurial repositories, and it doesn't work with some 3rd-party
providers of raw-file URLs like Codebase HQ's API.

This change converts our code to use the 'requests' module, which is downright
sexy. 'requests' will also do a much better job of trying to guess the charset
of any text we get.

Fixes bug 547
Reviewed at http://reviews.reviewboard.org/r/2863/