Remove all usage of ifuserorperm/ifnotuserandperm.
The ifuserorperm and ifnotuserandperm template tags appear to be pretty
broken with Django 1.4. I imagine this is due to their new code that
handles if/elif/endif, since it seems to trigger only when wrapped in an
if statement (in which case the opening tag is broken), or when an if
statement is wrapped inside of it (in which case the ending tag is
broken).

We don't really need these anymore, anyway. They were important in the
past because Django's if statement wouldn't do equality checks, so we
couldn't do an OR of the permission checks + user matching. Now we can,
so we can safely replace all these.

This fixes the review request page on Django 1.4.