Fix issues with repository management on Django 1.4.
All fields were being shown in the repository form, since Django's
templates changed the class names for field rows to be "field-{name}"
instead of just "{name}". This broke our showing/hiding logic.

Once you update some object in the database, the admin UI expects it can
send a message to the user's session, which requires the
MessagesMiddleware. This is now enabled.

And finally, the "Show SSH Key" was only showing up for existing
repositories, as the block it's in won't show up for new DB entries.
Move this out to a block that is always shown.