Better support moved files in the diff viewer.
Uses info from the diff to detect if files have been moved and displays
the result in the diff viewer.

This adds the functionality of parsing git's extended header of
detecting moved files. Before, the diff would show two changes: a
deleted file and a new file. Now, there is only one change and it is
displayed correctly in the diff viewer. Since a file may be moved and
have no actual diff data (and no revision), some extra checks are now
required to compensate for this. These may be different depending on the
SCM, so a more generic approach might be required.

Reviewed at http://reviews.reviewboard.org/r/2905/