Fix documentation building.
The recent change to use middleware for Review Board initialization
broke doc building. The MEDIA_SERIAL was never set, causing a very
confusing exception when trying to import reviewboard.reviews.views.

For future reference, the error was caused by MEDIA_SERIAL being used in
StarColumn in datagrids.py, but it wasn't yet defined. This was not
clear from the backtrace.

We now explicitly call initialize() to ensure we have everything we
need.