Switch our media setup to use Django's staticfiles support.
Django introduced built-in support for static media files in Django 1.3,
and have been improving upon this in 1.4. This staticfiles support makes
it possible for a site to specify where to look for media files (in app
directories, specific directories, or through custom Finder classes)
and, with a 'collectstatic' command, deploy them somewhere (another
directory or some other location through a Storage module, such as
Amazon S3). It also provides built-in support for post-processing
modules.

Django Pipeline now requires this support to even operate, and Django
1.4 is very strict about STATIC_ROOT not being empty and not being equal
to MEDIA_ROOT. So it's necessary that we move along with the world,
painful as it may be.

In this new world, there's a clear distinction between STATIC_* and
MEDIA_*. STATIC_* is used for media bundled with the application (our
own CSS, JavaScript, etc.), while MEDIA_* is for user-uploaded media
(screenshots, file attachments, etc.). This means that we can actually
in theory dump the files into different places (though maybe not yet in
practice).

One advantage this all does give us is that our initial setup is
simpler. Instead of prepare-dev having to get the djblets symlinks in
the right place, we can just rely on the staticfiles stuff to pull it in
from wherever it's installed. It also means that we can simplify our
unit tests' media management.

This will require a lot of testing and will likely break in some
circumstances.