Don't overwrite the entire environment when running collectstatic.
setup.py was overwriting the entire environment when attempting to set a
single environment variable in the call to collectstatic. This meant
that on the build machines, the virtualenv setup was being ignored, and
it was no longer able to find the modules needed.

We now make a copy of the environment, set the variable in that, and
then call collectstatic.