Don't allow manage.py to be run from within the reviewboard package dir.
manage.py now checks if it's being run from within the reviewboard
package directory (the same directory where manage.py lives). This is a
common problem that new developers hit, and can result in an SQLite
database ending up in the wrong location (due to relative paths in
settings_local.py). Now it just yells at you and tells you to run it in
the parent directory.