Support LDAP authentication without using anonymous bind credentials
This change modifies the search for user information so that when the Anonymous
credentials are not configured, the credentials entered by the user attempting
to log in are used to bind to the LDAP service in order to do the search.

I've tested this with our LDAP server, which is configured not to allow any
anonymous search; I believe that I did not change the behavior for the case in
which the anonymous credentials are configured (if I configure them, it does the
same sequence of operations to the LDAP server, but since they don't work I
cannot be 100% sure).

Fixes bug 2493
Reviewed at http://reviews.reviewboard.org/r/2909/