Disallow uploading diffs larger than 1 megabyte.
When diffs are particularly large, they can cause the server to spin at 100% CPU
for a very long time. This is normally not a problem (who changes that much code
at a time?), but parent diffs make it easier to hit.

This change fixes it pretty well for the upload forms on the web site.
post-review will also fail to upload, although the error message it prints out
is especially bad (it says that the diff file is empty).