Add original file and patched file resources to the Web API.
Adds resources for retrieving the original file, and the patched file
from the Web API.

These new resources are children of the File Diff Resource, and provide
a way for external tools to retrieve full copies of source files from a
review request.

If the file represented by the FileDiffResource is new, the link to the
OriginalFileResource will be removed. Alternatively, if the file is
deleted, the link to the PatchedFileResource will be removed.

Reviewed at http://reviews.reviewboard.org/r/3008/