Fix our media building setup.
Our media building worked okay with an existing development tree
(until recently -- will get to that in a second), but didn't work from a
fresh clone. Unfortunately, that broke our releases, as our release
script creates a fresh clone and builds from that.

The reason is that our media building would run ./manage.py collectstatic,
which required a settings_local.py file. That's dicey no matter what,
since we need certain settings to do this properly.

There's now a contrib/internal/build-media.py script, which sets us up
to use a specific settings_local.py file with the right settings and
paths. It can then safely build the media. Our setup.py script now calls
into this instead of directly running collectstatic.

Now, this basically works, except for the part where none of it does.
Django-Pipeline with Django 1.4 final has some problem with static
directories registered in STATICFILES_DIRS with prefixes ("rb",
"djblets", etc.). This is issue #95 on their bug tracker, and they have
a branch fixing the problem. As soon as that's released, we'll be able
to update to it and release 1.7 beta.