Fixed update notification popping up when it shouldn't.
There was a format mismatch with the template sending dates to
javascript, which compared it with a different date format and
thought there was a review update. Fixed this by sending the
datetime objects to javascript in ISO 8601 format, and wrapping
the dates in "localtime off" tags.

Also removed some unused imports.

Reviewed at http://reviews.reviewboard.org/r/3043/