Move public JavaScript functions into the RB namespace.
Pipeline now wraps all compressed JavaScript files in its own scope,
causing any functions defined within to be inaccessible outside the file
(such as to the <script> tags in the HTML). So now, we just stick these
in the RB namespace, like the API.

Down the road, we'll want to probably refactor these to be a bit more
object-oriented and better namespaced.