Fix the Review Board Activity admin widget with the new timestamp stuff.
The Review Board Activity widget was broken due to an assumption made
about the return type for the DateField on a given object. It was upset
that it wasn't getting a string on reviews.reviewboard.org, and was
instead getting a datetime.date. Now we cast to a string in order to get
the correct result.