Support getting GitHub API v2 tokens when configuring a repository.
GitHub is soon deprecating its older APIs in a switch to their v3 API.
As part of this, they have removed the API tokens from the accounts
page. However, they just today announced an API for getting these legacy
tokens.

This change adds a "Get your API token" button that prompts for a
password and then fetches the API token for the user and populates the
field.

With this, people should be able to set up new GitHub repositories
again.