Add is_authorized for determining if an account's authorization is valid.
HostingService.is_authorized returns whether or not an account is
authorized, based on the known data. This can be used when changing an
account's authorization method (for example, to change the hosting
service API that Review Board uses, or implementing one for a service
that doesn't yet use one) in order to flag that users need to
re-authorize.

GitHub uses this to make sure that the authorization data and token are
in the saved authorization data.