Fix all known remaining issues with hosting service defaults and saving.
Hosting service defaults should now be set correctly when loading an
existing repository and when switching between various settings on the
form.

It's a lot more flexible now with different types of hosting services.
It works fine whether or not the service supports different hosting plans,
and handles validation properly across all the sub-forms.