Make RB understand $Header$ keyword in Subversion
RB fails to display a diff ("The patch to 'foo.c' didn't apply cleanly.") if the
patch involves changes on the line containing $Header$ keyword. The reason is
that SVNTool does not know about the $Header$ keyword and thus does not collapse
it when preparing the original file for patching. $Header$ is standard SVN
keyword, see here:
http://svnbook.red-bean.com/en/1.6/svn.advanced.props.special.keywords.html

Reviewed at http://reviews.reviewboard.org/r/3104/