Don't cache failed manual updates checks.
Once a thread finds that manual updates are required, that state is
cached, requiring a restart to recompute the checks. This is
particularly problematic when a thread happens to hit the database when
it's temporarily down or something.

Now, we only cache if checks pass, and not when they fail. If they fail,
the next request will just try them again. That helps unstick these
threads.